# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Recherche d'emploi",
        "description": "Documentation de l'API Laravel pour la plateforme de recherche d'emploi",
        "contact": {
            "email": "hariznabil663@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Serveur principal"
        }
    ],
    "paths": {
        "/api/show-fans": {
            "get": {
                "tags": [
                    "CRUD _Users"
                ],
                "summary": "Afficher tous les utilisateurs avec leurs rôles",
                "description": "Retourne la liste des utilisateurs avec leurs rôles associés (accessible uniquement aux administrateurs).",
                "operationId": "079729b90e2da907adac9705f0ec84fe",
                "responses": {
                    "200": {
                        "description": "Liste des utilisateurs récupérée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "first_name": {
                                                "type": "string",
                                                "example": "Nabil"
                                            },
                                            "last_name": {
                                                "type": "string",
                                                "example": "Hariz"
                                            },
                                            "email": {
                                                "type": "string",
                                                "example": "nabil@test.ma"
                                            },
                                            "phone": {
                                                "type": "string",
                                                "example": "0612345678"
                                            },
                                            "roles": {
                                                "type": "array",
                                                "items": {
                                                    "type": "string",
                                                    "example": "admin"
                                                }
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non autorisé — Token ou rôle invalide"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/add-fan": {
            "post": {
                "tags": [
                    "CRUD _Users"
                ],
                "summary": "Créer un nouveau fan",
                "operationId": "29f33cf0fc3871dfa7a81484bf29980b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "first_name",
                                    "last_name",
                                    "email",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "first_name": {
                                        "type": "string",
                                        "example": "Nabil"
                                    },
                                    "last_name": {
                                        "type": "string",
                                        "example": "Hariz"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "nabil@test.ma"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "0612345678"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "password123"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "example": "password123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Fan créé avec succès"
                    },
                    "401": {
                        "description": "Non autorisé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/fan/{id}": {
            "put": {
                "tags": [
                    "CRUD _Users"
                ],
                "summary": "Mettre à jour un fan",
                "operationId": "3d25401571c9618757b8a5d7a206b808",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du fan à mettre à jour",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "first_name": {
                                        "type": "string",
                                        "example": "Nabil"
                                    },
                                    "last_name": {
                                        "type": "string",
                                        "example": "Hariz"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "nabil@test.ma"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "0612345678"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "newpassword123"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "example": "newpassword123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Fan mis à jour avec succès"
                    },
                    "404": {
                        "description": "Fan non trouvé"
                    },
                    "401": {
                        "description": "Non autorisé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "CRUD _Users"
                ],
                "summary": "Supprimer un fan",
                "operationId": "d3d3170de8b53b858c8fc866ce77336a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du fan à supprimer",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Fan supprimé avec succès"
                    },
                    "404": {
                        "description": "Fan non trouvé"
                    },
                    "401": {
                        "description": "Non autorisé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/fan-details/{id}": {
            "get": {
                "tags": [
                    "CRUD _Users"
                ],
                "summary": "Afficher le détail d'un fan",
                "operationId": "c49b9376ac378c4d3f3ce4311c9a3072",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du fan",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détails récupérés avec succès"
                    },
                    "404": {
                        "description": "Fan non trouvé"
                    },
                    "401": {
                        "description": "Non autorisé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Authentification"
                ],
                "summary": "connexion un utilisatuer ",
                "description": "Login user and create API token",
                "operationId": "296cf77f2ec37ef35b51e76f144be80f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "user@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "secret123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User logged in successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "user": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "first_name": {
                                                    "type": "string",
                                                    "example": "Nabil"
                                                },
                                                "last_name": {
                                                    "type": "string",
                                                    "example": "Hariz"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "example": "user@example.com"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "token": {
                                            "type": "string",
                                            "example": "1|Kf8kD6KJHfE3..."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Invalid credentials",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The provided credentials are incorrect."
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "Authentification"
                ],
                "summary": "Déconnexion de l'utilisateur",
                "operationId": "0ab4e3cf7bc4f7ec81b48f7928c7de7c",
                "responses": {
                    "200": {
                        "description": "Déconnexion réussie",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Logged out successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié (token manquant ou invalide)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/register": {
            "post": {
                "tags": [
                    "Authentification"
                ],
                "summary": "Register a new user",
                "operationId": "18235ea6963e661c49bf381065779d6a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "first_name",
                                    "last_name",
                                    "email",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "first_name": {
                                        "type": "string",
                                        "example": "Nabil"
                                    },
                                    "last_name": {
                                        "type": "string",
                                        "example": "Hariz"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "hariz@test.ma"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "0612345678"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "password"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "example": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User registered successfully"
                    }
                }
            }
        },
...
        "/api/contact": {
            "post": {
                "tags": [
                    "Contact"
                ],
                "summary": "Envoyer un message de contact",
                "operationId": "4d33e9ce28a28bf1184927750df24ba8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "message"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "john@example.com"
                                    },
                                    "message": {
                                        "type": "string",
                                        "example": "Bonjour, je veux plus d'infos."
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Message enregistré avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Message enregistré !"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "John Doe"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "example": "john@example.com"
                                                },
                                                "message": {
                                                    "type": "string",
                                                    "example": "Bonjour, je veux plus d'infos."
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2025-11-09T15:30:00Z"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2025-11-09T15:30:00Z"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erreur de validation"
                    }
                }
            }
        },
...
        "/api/match-details/{id}": {
            "get": {
                "tags": [
                    "Match"
                ],
                "summary": "Afficher le détail d'un match",
                "operationId": "b01ae00729e9ed41228d9492a311647f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du match",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Match récupéré avec succès"
                    },
                    "404": {
                        "description": "Match non trouvé"
                    },
                    "401": {
                        "description": "Non autorisé"
                    }
                }
            }
        },
...
        "/api/match/search": {
            "get": {
                "tags": [
                    "Match"
                ],
                "summary": "Rechercher des matchs",
                "operationId": "89a2ca3274a819924a9b01d0c918583d",
                "parameters": [
                    {
                        "name": "team",
                        "in": "query",
                        "description": "Nom de l'équipe",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "date",
                        "in": "query",
                        "description": "Date du match",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "city",
                        "in": "query",
                        "description": "Ville",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Résultat de la recherche"
                    },
                    "401": {
                        "description": "Non autorisé"
                    }
                }
            }
        },
...
        "/api/reservations/my": {
            "get": {
                "tags": [
                    "Reservation"
                ],
                "summary": "Afficher mes réservations (user)",
                "operationId": "15a5489bd77e1ceb1eadd132c949b69d",
                "responses": {
                    "200": {
                        "description": "Liste des réservations de l'utilisateur"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/stripe/webhook": {
            "post": {
                "tags": [
                    "Reservation"
                ],
                "summary": "Webhook Stripe pour créer des réservations après paiement",
                "operationId": "21a9524c0adbd35515c2e7e2825f77d4",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {}
                    }
                },
                "responses": {
                    "200": {
                        "description": "Webhook reçu avec succès"
                    },
                    "400": {
                        "description": "Erreur Stripe Webhook"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}